pyproject.toml
[tool.poetry]
name = "aerichtest"
version = "0.1.0"
description = ""
authors = ["ShirleyChau <2254751892@qq.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = "^0.16.15"
aerich = {git = "https://github.com/ShirleyChau/aerich.git"}

[[tool.poetry.source]]
name="tuna"
url="https://pypi.tuna.tsinghua.edu.cn/simple/"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
